+++ internal/provider/data_source_iosxe_pim_vrf_test.go
@@ -23,7 +23,10 @@ func TestAccDataSourceIosxePIMVRF(t *testing.T) {
 					resource.TestCheckResourceAttr("data.iosxe_pim_vrf.test", "bsr_candidate_priority", "10"),
 					resource.TestCheckResourceAttr("data.iosxe_pim_vrf.test", "bsr_candidate_accept_rp_candidate", "10"),
 					resource.TestCheckResourceAttr("data.iosxe_pim_vrf.test", "ssm_range", "10"),
-					resource.TestCheckResourceAttr("data.iosxe_pim_vrf.test", "ssm_default", "true"),
+					resource.TestCheckResourceAttr("data.iosxe_pim_vrf.test", "ssm_default", "false"),
+					resource.TestCheckResourceAttr("data.iosxe_pim_vrf.test", "rp_address", "19.19.19.19"),
+					resource.TestCheckResourceAttr("data.iosxe_pim_vrf.test", "rp_address_override", "false"),
+					resource.TestCheckResourceAttr("data.iosxe_pim_vrf.test", "rp_address_bidir", "true"),
 					resource.TestCheckResourceAttr("data.iosxe_pim_vrf.test", "rp_addresses.0.access_list", "10"),
 					resource.TestCheckResourceAttr("data.iosxe_pim_vrf.test", "rp_addresses.0.rp_address", "10.10.10.10"),
 					resource.TestCheckResourceAttr("data.iosxe_pim_vrf.test", "rp_addresses.0.override", "false"),
@@ -94,7 +97,10 @@ resource "iosxe_pim_vrf" "test" {
   bsr_candidate_priority = 10
   bsr_candidate_accept_rp_candidate = "10"
   ssm_range = "10"
-  ssm_default = true
+  ssm_default = false
+  rp_address = "19.19.19.19"
+  rp_address_override = false
+  rp_address_bidir = true
   rp_addresses = [{
     access_list = "10"
     rp_address = "10.10.10.10"